Automate clicking cookie button, continue to type text
Click the 'Reject All' button for the cookies, otherwise puppeteer will
start typing before the text box is available, making it actually start
typing at an offset (so then the wrong words are typed).

Also, monkeytype will stop taking input after a period of stable WPM, so
currently the bot keeps typing even on the results screen (since the
WPM was so stable that the test finished before all of the words were
typed). Still need to figure out how to resolve this.

Also need to check what's wrong with the viewport/how to change it
if needed.