Stop bot from typing after test ends
Previously, the bot continued typing after the test ended (to finish
the batch of words), so the end screen would continuously scroll to
the bottom. So you wouldn't see your result until the bot was finished
typing. Now, there is a while loop (line 43) to continuously check
that

1) the timer exists and
2) the time left isn't 0.

Only the first check is actually needed (otherwise, the result should
evaluate to null). Now I just need to work on improving the WPM to
letters per millisecond conversion (still inaccurate).