Clean code and comments in index.ts, populate README
Removed most debugging messages, added more comments in index.ts in
case I choose to redo or review this project later on. Added tips
on how to get started on recreating the project in the readme for
others (in particular, AWMies) to make their own typing bot.

A persistent issue is that the WPM is consistently off, which I tried
to remedy with line 43 in index.ts. The difference is more noticable
in higher WPMs (eg. 600), so it can be removed for more realistic
typing speeds like 100 WPM for better bot accuracy.

I'm satisfied with what I've learned and made with this project, so
this project is done for now. Hooray! :)